@@ -5,13 +5,13 @@ import { Order } from "./order.model";
 import { Observable } from "rxjs/Observable";
 
 import 'rxjs/add/operator/map';
-import { Http, RequestOptions, Headers } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 import { MEAT_API } from "app/app.api";
 
 @Injectable()
 export class OrderService {
     constructor(private cartService: ShoppingService,
-        private http: Http) { }
+        private http: HttpClient) { }
 
     cartItems(): CartItem[] {
         return this.cartService.items;
@@ -34,13 +34,7 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        const headers = new Headers();
-        headers.append('Content-Type', 'application/json');
-        return this.http.post(`${MEAT_API}/orders/`, JSON.stringify(order),
-            new RequestOptions({
-                headers: headers
-            }))
-            .map(response => response.json())
+        return this.http.post<Order>(`${MEAT_API}/orders/`, order)
             .map(order => order.id);
     }
 
